% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{auto_blank}
\alias{auto_blank}
\title{Automatically blank OD data.}
\usage{
auto_blank(data, forceBlanking, hasBlanks, blankData)
}
\arguments{
\item{data}{The tibble to analyze, as generated by \code{\link[grow96:processODData]{grow96::processODData()}}.}

\item{forceBlanking}{Logical value; if \code{TRUE}, blanking will use externally supplied blanks
regardless of whether the project has its own blank data.}

\item{hasBlanks}{Logical value from \code{auto_QC()}, indicating if the project has blank data.}

\item{blankData}{OD values for blanking. Can be a single value or a data frame of values.}
}
\value{
The original tibble, with an added column \code{blankedOD}.
}
\description{
A helper function used by \code{\link[=quick_analyze]{quick_analyze()}}.
This function calls \code{grow96::blankODs()} to calculate blanked ODs using either blanking data
from a project's own dataset, or supplied externally through the argument \code{blankData}.
}
\details{
When calling \code{grow96::blankODs()}, this function supplies the argument \code{method = "fixed"}
in order to calculate blanks using externally supplied values. See \code{\link[grow96:blankODs]{grow96::blankODs()}} for
more information on how blanking is implemented.
}
