% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{auto_process}
\alias{auto_process}
\title{Automatically process an OD dataset}
\usage{
auto_process(projectPath)
}
\arguments{
\item{projectPath}{The path of the project folder containing the spec and raw files
to be processed.}
}
\value{
A tidy tibble of all data from all experiments and replicates in the project.
}
\description{
A helper function used by \code{\link[=extract_blanks]{extract_blanks()}} and \code{\link[=quick_analyze]{quick_analyze()}}.
Using the supplied path for a folder containing a project's data files,
the function detects the prefix format for the data files and calls \code{\link[grow96:processODData]{grow96::processODData()}}
with the appropriate arguments.
}
