% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{auto_QC}
\alias{auto_QC}
\title{Automatically generate a quality control report for an OD dataset.}
\usage{
auto_QC(data, projectPath)
}
\arguments{
\item{data}{The tibble to analyze, as generated by \code{\link[grow96:processODData]{grow96::processODData()}}.}

\item{projectPath}{The path of the project folder containing the spec and raw files
to be processed.}
}
\value{
A logical value indicating if blanks were detected (\code{"TRUE"}) or not (\code{"FALSE"})
}
\description{
A helper function used by \code{\link[=quick_analyze]{quick_analyze()}}.
}
\details{
This function generates a quality control analysis for a project, saving it as a PDF
in the project's folder.

\code{auto_QC()} calls \code{\link[grow96:qcODData]{grow96::qcODData()}} and outputs a PDF of the report automatically.
Importantly, it also detects if a project has valid blanks based on the presence of data
in \verb{$qcBlanks}, which is then passed to the helper function \code{\link[=auto_blank]{auto_blank()}}.
}
\section{Notes}{
This function currently can't tell if valid blanks are missing from only some replicates
of an experiment. This is because \verb{$qcBlanks} will only be NULL if an entire experiment
fails to have blanks calculated for it.
}

