% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{extract_blanks}
\alias{extract_blanks}
\title{Extract blank values from grow96-compatible datasets.}
\usage{
extract_blanks(
  dataHubPath,
  colNames = NULL,
  variableKey = NULL,
  meansBy = NULL,
  combineData = FALSE,
  fileMethod = "save"
)
}
\arguments{
\item{dataHubPath}{The path of the "hub" folder containing project folders to analyze.}

\item{colNames}{A character vector of column names to check and fix if necessary. See Details
for more information.}

\item{variableKey}{A dataframe of definitions for replacing matching values in a single column.
See Details for more information.}

\item{meansBy}{A character vector of variables to group means by. If this argument is \code{NULL}
(the default), no mean blank ODs will be generated. See Details for more information.}

\item{combineData}{Defaults to \code{FALSE}. If \code{TRUE}, an additional dataframe called
"all_blanks" will be generated by combining dataframes from all projects. This argument will
automatically be \code{TRUE} if \code{meansBy} is defined.}
}
\value{
A list of tibbles of blank OD data. May include additional tibbles for combined blanks
and mean blank ODs depending on function arguments.
}
\description{
This function extracts blank values from any number of \code{grow96}-compatible datasets, with
options for calculating mean blank ODs and combining all blanks into a single large dataframe.
}
\details{
\code{extract_blanks()} is designed to generate blank values for use with \code{quick_analyze()}.
See the documentation for \code{\link[=quick_analyze]{quick_analyze()}} for information on how to organize project
datasets and how the output is formatted.
\subsection{Grouping means}{

Mean blank ODs are calculated by aggregating over all projects. Means can be grouped by
supplying a character vector to the argument \code{meansBy}. The vector must contain variables to
group by, which is then passed to \code{\link[=get_mean_blanks]{get_mean_blanks()}}.  \code{meansBy} is necessary for
generating blank values for use with \code{quick_analyze()} or \code{\link[grow96:blankODs]{grow96::blankODs}}.
}
}
\section{Automatically fixing columns/values}{

\subsection{Renaming column names}{

This function can rename slightly inconsistent variable names (such as \code{Drug} vs. \code{drugs}).
When \code{colNames} is a character vector of correct column names, it is passed to the helper
function \code{\link[=auto_rename]{auto_rename()}} and is used to check and fix incorrect column names. If column
joining causes an error somewhere, column names are likely responsible.
}

\subsection{Reassigning values in a column}{

This function can also replace values in a single column using a key with one column containing
the correct values and any number of additional columns containing project-specific values to
check for and replace. The first column of the key must be named identically to the data column
to check, and subsequent columns named identically to their corresponding project folder name.
\code{\link[=auto_rename]{auto_rename()}} will attempt to find cases in the data column that matches the project column
in the key, and replace it with the corresponding correct value from the first column of the key.
For safety, it also preserves the old values by copying it into a new column named in the format
\verb{<old name>_old}.
}
}

\examples{
extract_blanks(dataHubPath,
               colNames = NULL,
               variableKey = NULL,
               meansBy = NULL,
               combineData = FALSE,
               fileMethod = "save")

}
