% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_mean_blanks}
\alias{get_mean_blanks}
\title{Calculate mean ODs for blank data.}
\usage{
get_mean_blanks(data, meanGroups)
}
\arguments{
\item{data}{A dataframe/tibble of blanks to calculate the mean ODs of.}

\item{meanGroups}{A character vector of variable names to group the means by. Usually passed
from \code{extract_blanks()} through the argument \code{meansBy}.}
}
\value{
A tibble of meanODs with columns defined by \code{meanGroups}.
}
\description{
This function takes a dataframe/tibble of blank data and returns a tibble of mean ODs grouped
by the argument \code{meanGroups}.
}
\examples{
get_mean_blanks(data, meanGroups)

}
