% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{quick_analyze}
\alias{quick_analyze}
\title{Analyze grow96-compatible datasets.}
\usage{
quick_analyze(
  dataHubPath,
  forceBlanking = FALSE,
  blankData = NULL,
  colNames = NULL,
  exportGrowth = FALSE,
  combineData = FALSE,
  fileMethod = "save"
)
}
\arguments{
\item{dataHubPath}{The path of the "hub" folder containing project folders to analyze.}

\item{forceBlanking}{If \code{TRUE}, \code{blankData} is used for blanking regardless of whether
blank data exists in a particular project. Defaults to \code{FALSE}.}

\item{blankData}{A dataframe/tibble containing blank values. See \link[grow96:blankODs]{grow96::blankODs}
for more information.}

\item{colNames}{A character vector of column names to check and fix if necessary. See Details
for more information.}

\item{exportGrowth}{Defaults to \code{FALSE}. If \code{TRUE}, processed growth data will also be exported.}

\item{combineData}{Defaults to \code{FALSE}. If \code{TRUE}, an additional dataframe called "all_growths"
will be generated by combining processed growth data from all projects.}

\item{fileData}{Determines how the output is handled. The default value is \code{save}, which saves
the data as an RData file. To only return the data without saving it, use \code{return}. To do both,
use \code{both}.}
}
\value{
A list of \emph{n} elements, corresponding to \emph{n} projects, each containing a tibble of
processed and/or summarized OD data. May include an additional tibble of combined growth data.
}
\description{
\code{quick_analyze()} automatically analyzes any number of individual \code{grow96}-compatible datasets
and outputs processed and/or summarized growth data as dataframes in a list.
}
\details{
Each project's data files must be contained in a separate folder, and all project folders
placed in a single "hub" folder, the path for which is the argument \code{dataHubPath}.
By default, analyzed data is exported to the working directory, as an RData object with filename
\code{growthData_YYYY_MM_DD.RData}. The exported data is a list with as many elements as projects,
each element named according to project folder names. Only summarized data ("analyzed_data")
is exported by default, but unsummarized data ("processed_data") can also be exported by
setting \code{exportGrowth} to \code{TRUE}.

OD reads are blanked by calling \code{auto_blank()}. If necessary, blank values can be supplied
through the argument \code{blankData}, which is then used to blank OD data if a project does not
have valid blank data available. \code{forceBlanking} can be set to TRUE to force the use of
\code{blankData} regardless of whether blanks are available in a project.
\subsection{Renaming column names}{

This function can rename slightly inconsistent variable names (such as \code{Drug} vs. \code{drugs}).
When \code{colNames} is a character vector of correct column names, it is passed to the helper
function \code{\link[=auto_rename]{auto_rename()}} and is used to check and fix incorrect column names. If column
joining causes an error somewhere, column names are probably responsible somehow.
}
}
\examples{
quick_analyze(dataHubPath,
              forceBlanking = FALSE,
              blankData = NULL,
              colNames = NULL,
              exportGrowth = FALSE,
              combineData = FALSE,
              fileMethod = "save")

}
